requirements.txt
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
streamlit==1.30.0
scikit-learn==1.4.0